Refactor user ID generation in tracking and ingestion services
- Updated the Tracker class to use an empty string for customUserId instead of generating a new user ID.
- Enhanced the SessionReplayIngestService to generate a user ID server-side if not provided by the client, utilizing the userIdService.
- Replaced the user ID generation logic in createBasePayload with a call to userIdService for consistency.

Files modified:
- `server/src/analytics-script/tracking.ts`
- `server/src/services/replay/sessionReplayIngestService.ts`
- `server/src/tracker/utils.ts`
